Reorganise component and move into perf table
