+++ src/components/performance/PerfReport.tsx
@@ -22,7 +22,7 @@ import {
 } from '@blueprintjs/core';
 import { ItemPredicate, ItemRendererProps, Select } from '@blueprintjs/select';
 import { IconNames } from '@blueprintjs/icons';
-import { ColumnKeys, Columns, TypedPerfTableRow, getEligiblePerfColumns } from '../../definitions/PerfTable';
+import { ColumnKeys, Columns, TypedPerfTableRow } from '../../definitions/PerfTable';
 import { Signpost, calcHighDispatchOps } from '../../functions/perfFunctions';
 import SearchField from '../SearchField';
 import PerfTable from './PerfTable';
@@ -57,8 +57,6 @@ import { BufferType, BufferTypeLabel } from '../../model/BufferType';
 import { capitalizeString } from '../../functions/formatting';
 import { DeviceOperationLayoutTypes } from '../../model/APIData';
 import usePerfReportFiltering from './usePerfReportFiltering';
-import { useGetNPEManifest, useOpToPerfIdFiltered } from '../../hooks/useAPI';
-import PerfColumnPicker from './PerfColumnPicker';
 
 enum SignpostSelectType {
     START,
@@ -110,20 +108,6 @@ const PerformanceReport = ({
         () => Columns.filter((column) => column.filterable).map((column) => column.key),
         [],
     );
-    const opIdsMap = useOpToPerfIdFiltered();
-    const { data: npeManifest } = useGetNPEManifest();
-    const eligiblePerfColumns = useMemo(
-        () =>
-            getEligiblePerfColumns({
-                hasOpIds: opIdsMap.length > 0,
-                hasL1PressureData,
-                hiliteHighDispatch,
-                showHashColumn: false,
-                hasNpe: Boolean(npeManifest && npeManifest.length > 0),
-            }),
-        [opIdsMap.length, hasL1PressureData, hiliteHighDispatch, npeManifest],
-    );
-
     const [filters, setFilters] = useState(
         Object.fromEntries(filterableColumnKeys.map((key) => [key, ''] as [ColumnKeys, string])) as Record<
             ColumnKeys,
@@ -383,7 +367,7 @@ const PerformanceReport = ({
 
                 <div className='option-row'>
                     <FormGroup
-                        subLabel='Input data options'
+                        subLabel='Input data'
                         className='form-group'
                     >
                         <ButtonGroup className='select-group'>
@@ -603,19 +587,9 @@ const PerformanceReport = ({
                                         className='option-switch'
                                     />
                                 </Tooltip>
-
-                                {!isStackedView && <PerfColumnPicker eligibleColumns={eligiblePerfColumns} />}
                             </ButtonGroup>
                         </FormGroup>
                     )}
-
-                    {/* May keep this or remove it - undecided as yet */}
-                    {/* <Switch
-                        label='Show Hash Column'
-                        onChange={() => setShowHashColumn(!showHashColumn)}
-                        checked={showHashColumn}
-                        className='option-switch'
-                    /> */}
                 </div>
 
                 <Tabs
