+++ src/scss/components/PerfTableToolbar.scss
@@ -1,0 +1,27 @@
+// SPDX-License-Identifier: Apache-2.0
+//
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
+.perf-table-toolbar {
+    display: flex;
+    margin-bottom: 10px;
+
+    .perf-column-picker {
+        display: flex;
+        flex-direction: column;
+        padding: 8px;
+
+        .perf-column-picker-header {
+            display: flex;
+            align-items: center;
+            justify-content: space-between;
+            gap: 8px;
+            margin-bottom: 4px;
+        }
+
+        .column-selection {
+            margin-top: 4px;
+            margin-bottom: 4px;
+        }
+    }
+}
